ref-manual, poky.ent: Updated essential host packages
Fixes [YOCTO #13196]

The set of host packages for the testing has been merged into
the essential set of packages for all supported distros.  I
updated the variables defined in "poky.ent" so that these
packages are now part of the essentials.  I also removed the
section on the required packages for testing.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>