documentation: Cleaned up "plug-in"/"plugin" terminology.
The YP manual set was using the plugin term inconsistently.
It was appearing as "plugin" as well as "plug-in".  The
plugins in the project are all named without a hyphen.
The term itself is either/or according to Google.  I have
changed all references to plugin so that they do not use
a hyphen.

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>