sphinx: ref-variables: use builtin sphinx glossary
Sphinx has a glossary directive. From the documentation:
This directive must contain a reST definition list with terms and
definitions. The definitions will then be referencable with the 'term'
role.

So anywhere in *any* manual, we can do :term:`VAR` to refer to an item
from the glossary, and create a link.

To convert to a glossary, we needed proper indentation (e.g. added 3
spaces to each line)

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>